requirements.txt
asgiref==3.4.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-filter==21.1
django-log-viewer==1.1.4
django-phonenumber-field==5.2.0
django-rest-passwordreset==1.2.1
django-sendgrid-v5==1.1.1
django-url-filter==0.3.15
djangorestframework==3.12.4
enum-compat==0.0.3
idna==3.3
Markdown==3.3.4
phonenumbers==8.12.35
Pillow==8.4.0
Pygments==2.10.0
python-dotenv==0.19.1
python-http-client==3.3.3
pytz==2021.3
razorpay==1.2.0
requests==2.26.0
sendgrid==6.9.0
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
urllib3==1.26.7
